Add OSX version of CC3DemoMashUp demo app.
Add CC3DemoMashUp-Mac Xcode project and cocos3d-Mac workspace. Add -iOS to name of
existing demo projects and workspace. Replace all PVR textures in demos with PNG
for use in OSX. Refactor compiler switches for different OGL implementations.
Refactor CC3OpenGL class hierarchy. Add Mac support while using cocos2d 1.x.

+++ Demos/CC3DemoMashUp/CC3DemoMashUp/Classes/Controls/CCNodeAdornments.m
@@ -46,9 +46,9 @@
 
 @synthesize actionDuration;
 
--(int) zOrder { return [super zOrder]; }
+-(NSInteger) zOrder { return [super zOrder]; }
 
--(void) setZOrder: (int) z {
+-(void) setZOrder: (NSInteger) z {
 	CC2_ZORDER = z;
 }
